# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: mzt platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /releases:
    get:
      operationId: listReleases
      tags: [releases]
      summary: List all releases
      responses:
        "200":
          description: List of releases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Release"
    post:
      operationId: createRelease
      tags: [releases]
      summary: Create a release
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReleaseInput"
      responses:
        "201":
          description: Release created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /releases/{id}:
    get:
      operationId: getRelease
      tags: [releases]
      summary: Get a release with its tracks and reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Release detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReleaseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteRelease
      tags: [releases]
      summary: Delete a release
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      description: |
        Serves object entities uploaded via presigned URLs. These can optionally
        be protected with authentication or ACL checks based on the use case.
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir (e.g. `uploads/some-uuid`).
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - username
        - createdAt
...
    Track:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        audioUrl:
          type: string
        order:
          type: integer
      required:
        - id
        - title
        - audioUrl
        - order
    TrackInput:
      type: object
      properties:
        title:
          type: string
        audioUrl:
          type: string
      required:
        - title
        - audioUrl
    Release:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [single, album]
        artist:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        coverUrl:
          type: string
        audioUrl:
          type: ["string", "null"]
        isFromSanyaPlaylist:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          $ref: "#/components/schemas/User"
        averageScore:
          type: ["integer", "null"]
        reviewCount:
          type: integer
      required:
        - id
        - type
        - artist
        - title
        - coverUrl
        - createdAt
        - createdBy
        - reviewCount
    ReleaseDetail:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [single, album]
        artist:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        coverUrl:
          type: string
        audioUrl:
          type: ["string", "null"]
        isFromSanyaPlaylist:
          type: boolean
        createdAt:
          type: string
          format: date-time
        createdBy:
          $ref: "#/components/schemas/User"
        averageScore:
          type: ["integer", "null"]
        reviewCount:
          type: integer
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/Track"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        userReview:
          oneOf:
            - $ref: "#/components/schemas/Review"
            - type: "null"
      required:
        - id
        - type
        - artist
        - title
        - coverUrl
        - createdAt
        - createdBy
        - reviewCount
        - tracks
        - reviews
    ReleaseInput:
      type: object
      properties:
        type:
          type: string
          enum: [single, album]
        artist:
          type: string
          minLength: 1
        title:
          type: string
          minLength: 1
        description:
          type: string
        coverUrl:
          type: string
        audioUrl:
          type: string
        tracks:
          type: array
          items:
            $ref: "#/components/schemas/TrackInput"
      required:
        - type
        - artist
        - title
        - coverUrl
    Review:
      type: object
      properties:
        id:
          type: integer
        releaseId:
          type: integer
        user:
          $ref: "#/components/schemas/User"
        rhymes:
          type: integer
          minimum: 1
          maximum: 10
        structure:
          type: integer
          minimum: 1
          maximum: 10
        styleExecution:
          type: integer
          minimum: 1
          maximum: 10
        individuality:
          type: integer
          minimum: 1
          maximum: 10
        atmosphere:
          type: integer
          minimum: 1
          maximum: 10
        score:
          type: integer
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - releaseId
        - user
        - rhymes
        - structure
        - styleExecution
        - individuality
        - atmosphere
        - score
        - comment
        - createdAt
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file (e.g. `image/jpeg`).
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path (e.g. `/objects/uploads/uuid`). Store this in your database.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'