# lib/api-zod/src/generated/types/release.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * mzt platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReleaseType } from './releaseType';
import type { User } from './user';

export interface Release {
  id: number;
  type: ReleaseType;
  artist: string;
  title: string;
  /** @nullable */
  description?: string | null;
  coverUrl: string;
  /** @nullable */
  audioUrl?: string | null;
  isFromSanyaPlaylist?: boolean;
  createdAt: Date;
  createdBy: User;
  /** @nullable */
  averageScore?: number | null;
  reviewCount: number;
}