# lib/api-zod/src/generated/types/releaseDetail.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * mzt platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ReleaseDetailType } from './releaseDetailType';
import type { Review } from './review';
import type { Track } from './track';
import type { User } from './user';

export interface ReleaseDetail {
  id: number;
  type: ReleaseDetailType;
  artist: string;
  title: string;
  /** @nullable */
  description?: string | null;
  coverUrl: string;
  /** @nullable */
  audioUrl?: string | null;
  isFromSanyaPlaylist?: boolean;
  createdAt: Date;
  createdBy: User;
  /** @nullable */
  averageScore?: number | null;
  reviewCount: number;
  tracks: Track[];
  reviews: Review[];
  userReview?: Review | null;
}